DayOffUpdate 추가 및 사원 정보 print시 float 정상 출력되도록 수정

+++ src/com/konkuk/controller/EmployeeController.java
@@ -64,7 +64,7 @@ public class EmployeeController extends Controller {
                 exactfind.forEach(employee -> {
                     UI.print("*" + Langs.EMPLOYEE_ID + employee.id);
                     UI.print("*" + Langs.EMPLOYEE_NAME + employee.name);
-                    UI.print("*" + Langs.EMPLOYEE_SALARY + employee.salary + "\n");
+                    UI.print("*" + Langs.EMPLOYEE_SALARY + (int) employee.salary + "\n");
                 });
                 break;
             }
@@ -98,7 +98,7 @@ public class EmployeeController extends Controller {
                 if(Integer.parseInt(id) == Erepositry.findByExactId(Integer.parseInt(id)).id){
                     UI.print("*"+Langs.EMPLOYEE_ID+Erepositry.findByExactId(Integer.parseInt(id)).id);
                     UI.print("*"+Langs.EMPLOYEE_NAME+Erepositry.findByExactId(Integer.parseInt(id)).name);
-                    UI.print("*"+Langs.EMPLOYEE_SALARY+Erepositry.findByExactId(Integer.parseInt(id)).salary+"\n");
+                    UI.print("*"+Langs.EMPLOYEE_SALARY+(int) Erepositry.findByExactId(Integer.parseInt(id)).salary+"\n");
                 }else { //검색대상 오류
                     // 만약 올바르지 않은 숫자거나 문자라면
                     UI.print(Langs.LETTER_ERROR);
